feat: application router and controller created, index router updated
