+++ backend/src/routers/animalRouter.js
@@ -4,10 +4,10 @@ import { isAuth } from "../middlewares/authentication.middleware.js";
 
 export const router = Router();
 
-router.get("/animals", animalController.getAllAnimals);
-router.get("/animals/:id", animalController.getOneAnimal);
-router.get("/animals/species", animalController.getSpecies);
+router.get("/", animalController.getAllAnimals);
+router.get("/:id", animalController.getOneAnimal);
+router.get("/species", animalController.getSpecies);
 
-router.post("/animals", isAuth, animalController.createAnimal);
-router.put("/animals/:id", isAuth, animalController.updateAnimal);
-router.delete("/animals/:id", isAuth, animalController.deleteAnimal);
+router.post("/", isAuth, animalController.createAnimal);
+router.put("/:id", isAuth, animalController.updateAnimal);
+router.delete("/:id", isAuth, animalController.deleteAnimal);
