+++ backend/src/routers/index.js
@@ -1,12 +1,10 @@
 import { Router } from "express";
 import { router as userRouter } from "./userRouter.js";
 import { router as animalRouter } from "./animalRouter.js";
-import { router as associationRouter } from "./associationRouter.js";
+import { router as applicationRouter } from "./applicationRouter.js";
 
 export const router = Router();
 
 router.use("/users",userRouter);
-
-router.use(animalRouter);
-
-router.use(associationRouter);
+router.use("/animals", animalRouter);
+router.use("/applications", applicationRouter);
